modified:   app/Http/Controllers/EmployeeController.php 	modified:   app/Http/Controllers/UserController.php

+++ app/Http/Controllers/EmployeeController.php
@@ -8,9 +8,8 @@ use App\Models\EmployeePost;
 use App\Models\Operator;
 use App\Models\Status;
 use App\Models\User;
-use App\Models\UserRole;
-use App\Models\UserStatus;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Validation\Rule;
 
@@ -31,10 +30,10 @@ class EmployeeController extends Controller
                             ->orWhere('last_name', 'like', "%{$search}%");
                     });
             });
-            if ($request->department != null) { 
+            if ($request->department != null) {
                 $query->where('department_id', $request->department);
             }
-    
+
             if ($request->status != null) {
                 $query->where('status_id', $request->status);
             }
@@ -54,83 +53,92 @@ class EmployeeController extends Controller
         $operators = Operator::all();
         $posts = EmployeePost::all();
 
-        return view('employees.create', compact('departments', 'operators','posts'));
+        return view('employees.create', compact('departments', 'operators', 'posts'));
     }
 
     public function store(Request $request)
     {
-        $request->validate([
-            'first_name' => 'required|string|max:255',
-            'last_name' => 'required|string|max:255',
-            'email' => 'required|email|max:255|unique:users',
-            'password' => 'required|min:8',
-
-            'department_id' => 'required|exists:departments,id',
-            'profile_picture' => ['required','image','mimes:jpeg,png,jpg','dimensions:width=1080,height=1080,ratio=1/1'],
-            'personal_num' => 'required|string|max:255',
-            'salary' => 'required|numeric',
-            'address' => 'required|string|max:255',
-            'cin' => 'required|string|max:8|unique:employees',
-        ]);
-
-        $user = User::create([
-            'first_name' => $request->first_name,
-            'last_name' => $request->last_name,
-            'email' => $request->email,
-            'password' => $request->password,
-            'user_role_id' => 3,
-        ]);
-
-        if ($request->hasFile('profile_picture')) {
-            $path = $request->file('profile_picture')->store('profile_pictures', 'public');
-            // $validated['profile_picture'] = $path;
-            $pic_path = $path;
-        }
-
-        $data = [
-            'department_id' => $request->department_id,
-            'profile_picture' => $pic_path,
-            'personal_num' => $request->personal_num,
-            'salary' => $request->salary,
-            'address' => $request->address,
-            'status_id' => 1,
-            'user_id' => $user->id,
-            'cin' => $request->cin
-        ];
-
-        if ($request->is_freelancer === 'freelancer') {
+        DB::beginTransaction();
+        try {
             $request->validate([
-                'ice' => 'required|string|max:255',
+                'first_name' => 'required|string|max:255',
+                'last_name' => 'required|string|max:255',
+                'email' => 'required|email|max:255|unique:users',
+                'password' => 'required|min:8',
+
+                'department_id' => 'required|exists:departments,id',
+                'profile_picture' => ['required', 'image', 'mimes:jpeg,png,jpg', 'dimensions:width=1080,height=1080,ratio=1/1'],
+                'personal_num' => 'required|string|max:255',
+                'salary' => 'required|numeric',
+                'address' => 'required|string|max:255',
+                'cin' => 'required|string|max:8|unique:employees',
             ]);
 
-            $data['ice'] = $request->ice;
-            $data['is_project'] = $request->is_project ? true : false;
-            $data['hours'] = 0;
-        } else {
-            $request->validate([
-                'employee_code' => 'required|string|max:255|unique:employees',
-                'professional_num' => 'required|string|max:255',
-                'pin' => 'required|string|max:255',
-                'puk' => 'required|string|max:255',
-                'operator_id' => 'required|exists:operators,id',
-                'professional_email' => 'required|email|max:255',
-                'cnss' => 'required|string|max:255',
-                'assurance' => 'required|string|max:255',
+            $user = User::create([
+                'first_name' => $request->first_name,
+                'last_name' => $request->last_name,
+                'email' => $request->email,
+                'password' => bcrypt($request->password),
+                'user_role_id' => 3,
             ]);
 
-            $data['employee_code'] = $request->employee_code;
-            $data['professional_num'] = $request->professional_num;
-            $data['professional_email'] = $request->professional_email;
-            $data['pin'] = $request->pin;
-            $data['puk'] = $request->puk;
-            $data['operator_id'] = $request->operator_id;
-            $data['cnss'] = $request->cnss;
-            $data['assurance'] = $request->assurance;
-        }
-        Employee::create($data);
+            $pic_path = null;
 
-        return redirect()->route('employees.index')
-            ->with('success', 'Employee created successfully.');
+            if ($request->hasFile('profile_picture')) {
+                $path = $request->file('profile_picture')->store('profile_pictures', 'public');
+                // $validated['profile_picture'] = $path;
+                $pic_path = $path;
+            }
+
+            $data = [
+                'department_id' => $request->department_id,
+                'profile_picture' => $pic_path,
+                'personal_num' => $request->personal_num,
+                'salary' => $request->salary,
+                'address' => $request->address,
+                'status_id' => 1,
+                'user_id' => $user->id,
+                'cin' => $request->cin
+            ];
+
+            if ($request->is_freelancer === 'freelancer') {
+                $request->validate([
+                    'ice' => 'required|string|max:255',
+                ]);
+
+                $data['ice'] = $request->ice;
+                $data['is_project'] = $request->is_project ? true : false;
+                $data['hours'] = 0;
+            } else {
+                $request->validate([
+                    'employee_code' => 'required|string|max:255|unique:employees',
+                    'professional_num' => 'required|string|max:255',
+                    'pin' => 'required|string|max:255',
+                    'puk' => 'required|string|max:255',
+                    'operator_id' => 'required|exists:operators,id',
+                    'professional_email' => 'required|email|max:255',
+                    'cnss' => 'required|string|max:255',
+                    'assurance' => 'required|string|max:255',
+                ]);
+
+                $data['employee_code'] = $request->employee_code;
+                $data['professional_num'] = $request->professional_num;
+                $data['professional_email'] = $request->professional_email;
+                $data['pin'] = $request->pin;
+                $data['puk'] = $request->puk;
+                $data['operator_id'] = $request->operator_id;
+                $data['cnss'] = $request->cnss;
+                $data['assurance'] = $request->assurance;
+            }
+            Employee::create($data);
+
+            return redirect()->route('employees.index')
+                ->with('success', 'Employee created successfully.');
+        } catch (\Exception $e) {
+            DB::rollback();
+
+            return back()->withErrors(['error' => 'Something went wrong: ' . $e->getMessage()]);
+        }
     }
 
     public function show(Employee $employee)
@@ -146,7 +154,7 @@ class EmployeeController extends Controller
         $operators = Operator::all();
         $posts = EmployeePost::all();
 
-        return view('employees.edit', compact('employee','departments', 'operators','posts'));
+        return view('employees.edit', compact('employee', 'departments', 'operators', 'posts'));
     }
 
     public function update(Request $request, Employee $employee)
@@ -189,11 +197,14 @@ class EmployeeController extends Controller
 
     public function destroy(Employee $employee)
     {
-        if ($employee->profile_picture) {
-            Storage::disk('public')->delete($employee->profile_picture);
-        }
+        DB::transaction(function () use ($employee) {
+            if ($employee->profile_picture) {
+                Storage::disk('public')->delete($employee->profile_picture);
+            }
 
-        $employee->delete();
+            User::where('user_id', $employee->id)->delete();
+            $employee->delete();
+        });
 
         return redirect()->route('employees.index')
             ->with('success', 'Employee deleted successfully.');
